+++ package.json
@@ -69,13 +69,11 @@
     "@observablehq/plot": "github:chaichontat/plot",
     "colormap": "^2.3.2",
     "d3": "^7.6.1",
-    "d3-axis": "^3.0.0",
     "fzf": "^0.5.1",
     "lodash-es": "^4.17.21",
     "mdsvex-relative-images": "^1.0.3",
     "pako": "^2.1.0",
     "papaparse": "^5.3.2",
-    "tippy.js": "^6.3.7",
-    "vega-embed": "^6.21.0"
+    "tippy.js": "^6.3.7"
   }
 }
