+++ src/lib/components/graph.svelte
@@ -1,4 +1,4 @@
-<script lang="ts">
+<!-- <script lang="ts">
   import { onMount } from "svelte";
   import embed, { type VisualizationSpec } from "vega-embed";
   export let spec: VisualizationSpec;
@@ -14,4 +14,4 @@
 <aside class="inline-block"><slot /></aside>
 <div class="not-prose flex w-full justify-evenly">
   <figure id={`vis-${uid}`} />
-</div>
+</div> -->
