--- src/routes/previews/[slug]/+page.svelte
@@ -1,8 +1,0 @@
-<script lang="ts">
-  import Preview from "$src/lib/elements/preview.svelte";
-  import type { PageData } from "./$types";
-
-  export let data: PageData;
-</script>
-
-<Preview post={data.post} />
