+++ src/routes/tags/[slug]/+page.ts
@@ -1,12 +1,13 @@
-import { posts } from "$src/lib/slug";
+import { posts } from "$lib/slug";
 import { error } from "@sveltejs/kit";
 import type { PageLoad } from "./$types";
 
 export const load: PageLoad = ({ params }) => {
-  const filteredPost = posts.find((post) => post.slug === params.slug.toLowerCase());
+  const filteredPost = posts.filter((post) => post.post.metadata.tags?.includes(params.slug));
   // const header = url.searchParams.get("header") === "false" ? false : true;
   if (!filteredPost) {
     throw error(404, "Not found.");
   }
-  return { post: filteredPost };
+
+  return { filtered: filteredPost, slug: params.slug };
 };
