--- src/routes/tags/__layout@pad.svelte
@@ -1,1 +1,0 @@
-<slot />
